+++ packages/components/input/src/Input.tsx
@@ -66,7 +66,6 @@ const Root = forwardRef<HTMLInputElement, InputProps>(
         className={inputStyles({
           className,
           intent: state,
-          isStandalone: !!group.isStandalone,
           hasLeadingAddon: !!hasLeadingAddon,
           hasTrailingAddon: !!hasTrailingAddon,
           hasLeadingIcon: !!hasLeadingIcon,