+++ packages/components/textarea/src/Textarea.stories.tsx
@@ -64,7 +64,7 @@ export const GroupDisabled: StoryFn = _args => (
 
 export const FieldLabel: StoryFn = _args => {
   return (
-    <FormField className="!gap-sm" name="title">
+    <FormField name="title">
       <FormField.Label>Title</FormField.Label>
 
       <Textarea rows={2} />
@@ -74,7 +74,7 @@ export const FieldLabel: StoryFn = _args => {
 
 export const FieldHiddenLabel: StoryFn = _args => {
   return (
-    <FormField className="!gap-sm" name="title">
+    <FormField name="title">
       <FormField.Label>
         <VisuallyHidden>Title</VisuallyHidden>
       </FormField.Label>
@@ -86,7 +86,7 @@ export const FieldHiddenLabel: StoryFn = _args => {
 
 export const FieldRequired: StoryFn = _args => {
   return (
-    <FormField className="!gap-sm" name="title" isRequired>
+    <FormField name="title" isRequired>
       <FormField.Label>Title</FormField.Label>
 
       <Textarea rows={2} />
@@ -96,7 +96,7 @@ export const FieldRequired: StoryFn = _args => {
 
 export const FieldHelperMessage: StoryFn = _args => {
   return (
-    <FormField className="!gap-sm" name="title">
+    <FormField name="title">
       <FormField.Label>Title</FormField.Label>
 
       <Textarea rows={2} />
@@ -108,9 +108,37 @@ export const FieldHelperMessage: StoryFn = _args => {
   )
 }
 
+export const FieldCharactersCount: StoryFn = _args => {
+  const MAX_LENGTH = 90
+  const [value, setValue] = useState('')
+
+  const handleChange = (event: ChangeEvent<HTMLTextAreaElement>) => {
+    setValue(event.target.value)
+  }
+
+  return (
+    <FormField name="textarea-with-a-characters-count">
+      <FormField.Label>Textarea with a characters count</FormField.Label>
+
+      <Textarea value={value} onChange={handleChange} maxLength={MAX_LENGTH} />
+
+      <div className="flex justify-between gap-md">
+        <div className="grow">
+          <FormField.HelperMessage>
+            Type the text but take into account the max length
+          </FormField.HelperMessage>
+          <FormField.ErrorMessage>This is an error</FormField.ErrorMessage>
+        </div>
+
+        <FormField.CharactersCount value={value} maxLength={MAX_LENGTH} />
+      </div>
+    </FormField>
+  )
+}
+
 export const FieldInvalid: StoryFn = _args => {
   return (
-    <FormField className="!gap-sm" name="title" state="error">
+    <FormField name="title" state="error">
       <FormField.Label>Title</FormField.Label>
 
       <Textarea rows={2} />