+++ packages/components/radio-group/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.9](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.3.8...@spark-ui/radio-group@1.3.9) (2023-07-19)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
+## [1.3.8](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.3.7...@spark-ui/radio-group@1.3.8) (2023-07-19)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
 ## [1.3.7](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.3.6...@spark-ui/radio-group@1.3.7) (2023-07-17)
 
 **Note:** Version bump only for package @spark-ui/radio-group